+++ test/github-issues/144/issue-144.ts
@@ -3,7 +3,7 @@ import {createTestingConnections, closeTestingConnections, reloadTestingDatabase
 import {Connection} from "../../../src/connection/Connection";
 import {Student} from "./entity/Student";
 
-describe.only("github issues > #144 Class Table Inheritance doesn't seem to work", () => {
+describe("github issues > #144 Class Table Inheritance doesn't seem to work", () => {
 
     let connections: Connection[];
     before(async () => connections = await createTestingConnections({